Mbot/requirements.txt
appdirs==1.4.3
diff-match-patch==20121119
# VERSION DJANGO
Django==1.10.5

# DJANGO BOOTSTRAP
django-bootstrap3==8.2.2

# DJANGO IMPORT EXPORT
django-import-export==0.5.1

# DJANGO FUTURE
future==0.16.0

# Json Database
jsondatabase==0.1.7

oauthlib==2.0.1
olefile==0.44
packaging==16.8

# Gestor de imagenes
Pillow==4.0.0

# pkg-resources==0.0.0

# Conexion con base de datos postgres
psycopg2==2.7.1

# QR CODE
# qrcode-5.3

pyparsing==2.2.0
python-twitter==3.2.1
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
tablib==0.11.4

# Para hacer las imagenes mas pequenias
	django-imagekit==4.0

# Para el calendario
	django-datetime-widget==0.9.3

# Reporlab
# reportlab==3.4.0

# ordenar los requerimientos

# Django suit admin
# django-suit==0.2.25


# Install mitmproxy
sudo pip install mitmproxy

appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
decorator==4.0.11
diff-match-patch==20121119

django-appconf==1.0.2
django-bootstrap3==8.2.2

#Django Constance
django-constance==2.0.0

#Django ImagenKit
django-imagekit==4.0

# DJANGO REDIS
django-redis==4.6.0

#Install Fabric 
Fabric==1.13.1

future==0.16.0
idna==2.5

# Django ipython
ipython==5.3.0

ipython-genutils==0.2.0
jsondatabase==0.1.7
oauthlib==2.0.1
olefile==0.44
packaging==16.8
paramiko==2.1.2
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.0.0

prompt-toolkit==1.0.14
#Connection Postgresql
psycopg2==2.7.1

ptyprocess==0.5.1
pyasn1==0.2.3
pycparser==2.17

#Install Redis
redis==2.10.5

# Reportes PDF
pip install xhtml2pdf --pre

# Reportes Excel
openpyxl==2.4.9

requests==2.12.4
requests-oauthlib==0.7.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
tablib==0.11.4
traitlets==4.3.2

#Django Transitions
transitions==0.4.2
wcwidth==0.1.7


#Django restFramework

djangorestframework==3.7.1